@@ -14,9 +14,9 @@
 import copy
 from copy import deepcopy
 from dataclasses import dataclass, fields
-from typing import Generator, List, Optional
+from typing import Optional, Union
 
-__all__ = ["QuantizationScheme", "preset_name_to_scheme"]
+__all__ = ["QuantizationScheme", "get_gguf_scheme", "preset_name_to_scheme"]
 
 
 @dataclass
@@ -32,14 +32,13 @@ class QuantizationScheme:
     act_dynamic: Optional[bool] = None
     super_bits: Optional[int] = None
     super_group_size: Optional[int] = None
-    clip: Optional[bool] = False
 
     @classmethod
     def from_dict(cls, config: dict):
         return cls(**config)
 
     @classmethod
-    def get_attributes(cls: "QuantizationScheme") -> List[str]:
+    def get_attributes(cls: "QuantizationScheme") -> list[str]:
         return [field.name for field in fields(cls)]
 
     def __getitem__(self, key: str):
@@ -73,7 +72,15 @@ class QuantizationScheme:
     def __eq__(self, other: "QuantizationScheme") -> bool:
         if not isinstance(other, QuantizationScheme):
             return False
+        skip_act_check = False
+        self_act_bits = 16 if self.act_bits is None else self.act_bits
+        other_act_bits = 16 if other.act_bits is None else other.act_bits
+        if self_act_bits == other_act_bits and other_act_bits >= 16:
+            skip_act_check = True
+
         for field in self.get_attributes():
+            if skip_act_check and field.startswith("act_"):
+                continue
             if getattr(self, field) != getattr(other, field):
                 return False
         return True
@@ -115,6 +122,27 @@ W2A16 = QuantizationScheme.from_dict(
     }
 )
 
+W2A16G64 = QuantizationScheme.from_dict(
+    {
+        "bits": 2,
+        "sym": True,
+        "group_size": 64,
+        "data_type": "int",
+        "act_bits": 16,
+    }
+)
+
+
+W2A16G32 = QuantizationScheme.from_dict(
+    {
+        "bits": 2,
+        "sym": True,
+        "group_size": 32,
+        "data_type": "int",
+        "act_bits": 16,
+    }
+)
+
 W3A16 = QuantizationScheme.from_dict(
     {
         "bits": 3,
@@ -143,6 +171,7 @@ MXFP4 = QuantizationScheme.from_dict(
         "act_bits": 4,
         "act_data_type": "mx_fp_rceil",
         "act_group_size": 32,
+        "act_sym": True,
     }
 )
 
@@ -154,6 +183,7 @@ MXFP8 = QuantizationScheme.from_dict(
         "act_bits": 8,
         "act_data_type": "mx_fp_rceil",
         "act_group_size": 32,
+        "act_sym": True,
     }
 )
 
@@ -165,6 +195,7 @@ NVFP4 = QuantizationScheme.from_dict(
         "act_bits": 4,
         "act_data_type": "nv_fp4_with_static_gs",
         "act_group_size": 16,
+        "act_sym": True,
     }
 )
 
@@ -177,6 +208,7 @@ FPW8A16 = QuantizationScheme.from_dict(
         "act_data_type": "fp",
     }
 )
+
 
 # FP8 = asdict(QuantArgs.from_dict({
 #     "bits": 8,
@@ -199,6 +231,17 @@ FP8_STATIC = QuantizationScheme.from_dict(
     }
 )
 
+# For AutoScheme 16 bits options
+BF16 = QuantizationScheme.from_dict(
+    {
+        "bits": 16,
+        "group_size": 128,
+        "data_type": "fp",
+        "act_bits": 16,
+        "act_data_type": "fp",
+    }
+)
+
 PRESET_SCHEMES = {
     "W4A16": W4A16,
     "W2A16": W2A16,
@@ -208,7 +251,10 @@ PRESET_SCHEMES = {
     "MXFP8": MXFP8,
     "NVFP4": NVFP4,
     "FPW8A16": FPW8A16,
+    "W2A16G64": W2A16G64,
+    "W2A16G32": W2A16G32,
     "FP8_STATIC": FP8_STATIC,
+    "BF16": BF16,
 }
 from auto_round.export.export_to_gguf.config import GGUF_CONFIG
 
@@ -218,3 +264,22 @@ for key, val in GGUF_CONFIG.items():
     value.pop("embedding", None)
     value.pop("lm_head", None)
     PRESET_SCHEMES[key.upper()] = QuantizationScheme.from_dict(value)
+
+
+def get_gguf_scheme(scheme: Union[str, QuantizationScheme]) -> str:
+    if isinstance(scheme, str) and scheme.upper().startswith("GGUF"):
+        return scheme
+    if isinstance(scheme, str):
+        return ""
+    for key, val in PRESET_SCHEMES.items():
+        # For q40 or q4_1 we only support it with str scheme， otherwise it will be matched incorrectly with W4G32
+        if not key.upper().startswith("GGUF") or ("0" in key or "1" in key):
+            continue
+        equal = True
+        for scheme_key in val.keys():
+            if val[scheme_key] is not None and val[scheme_key] != scheme.get(scheme_key, None):
+                equal = False
+                break
+        if equal:
+            return key
+    return ""
