@@ -11,8 +11,12 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-from auto_round.autoround import AutoRound, AutoRoundAdam
-from auto_round.mllm import AutoRoundMLLM
+from auto_round.autoround import AutoRound
+
+# support for old api
+from auto_round.autoround import AutoRoundLLM, AutoRoundMLLM, AutoRoundAdam, AutoRoundDiffusion
+from auto_round.schemes import QuantizationScheme
+from auto_round.auto_scheme import AutoScheme
 from auto_round.utils import LazyImport
 
 
