@@ -21,7 +21,8 @@ import torch
 import transformers
 
 from auto_round.export.register import register_format
-from auto_round.utils import check_to_quantized, detect_device, get_module, logger, set_module
+from auto_round.logger import logger
+from auto_round.utils import check_to_quantized, detect_device, get_module, set_module
 
 from .config import QuantConfig
 from .model_wrapper import WeightOnlyLinear
@@ -226,13 +227,13 @@ def pack_model(
             else:
                 scale = scale.to(dtype=convert_dtype)
                 zp = zp.to(dtype=torch.int32) if isinstance(zp, torch.Tensor) else zp
-        if isinstance(m, transformers.pytorch_utils.Conv1D):
+        if type(m) == transformers.pytorch_utils.Conv1D:
             fp_weight = fp_weight.t_().contiguous()
         int_weight = quant_weight_w_scale(fp_weight, scale, zp, group_size, fp_weight.device)
-        if isinstance(m, torch.nn.Linear):
+        if type(m) == torch.nn.Linear:
             in_features = m.in_features
             out_features = m.out_features
-        elif isinstance(m, transformers.pytorch_utils.Conv1D):
+        elif type(m) == transformers.pytorch_utils.Conv1D:
             in_features = m.weight.shape[0]
             out_features = m.weight.shape[1]
         int_weight = int_weight.type(torch.int32)
