@@ -22,7 +22,7 @@ class LLMDataLoader:
 class TestAutoRoundFormatGeneration(unittest.TestCase):
     @classmethod
     def setUpClass(self):
-        self.model_name = "facebook/opt-125m"
+        self.model_name = "/tf_dataset/auto_round/models/facebook/opt-125m"
         self.model = AutoModelForCausalLM.from_pretrained(self.model_name, torch_dtype="auto", trust_remote_code=True)
         self.tokenizer = AutoTokenizer.from_pretrained(self.model_name, trust_remote_code=True)
         self.llm_dataloader = LLMDataLoader()
