@@ -12,6 +12,7 @@ from auto_round import AutoRound
 
 
 class LLMDataLoader:
+
     def __init__(self):
         self.batch_size = 1
 
@@ -21,10 +22,10 @@ class LLMDataLoader:
 
 
 class TestGGUF(unittest.TestCase):
+
     @classmethod
     def setUpClass(self):
-        self.model_name = "Qwen/Qwen2.5-0.5B-Instruct"
-        self.model = AutoModelForCausalLM.from_pretrained(self.model_name, torch_dtype="auto", trust_remote_code=True)
+        self.model_name = "/tf_dataset/auto_round/models/Qwen/Qwen2.5-0.5B-Instruct"
         self.tokenizer = AutoTokenizer.from_pretrained(self.model_name, trust_remote_code=True)
         self.llm_dataloader = LLMDataLoader()
 
@@ -36,7 +37,7 @@ class TestGGUF(unittest.TestCase):
     def test_basic_usage(self):
         python_path = sys.executable
         res = os.system(
-            f"cd ../.. && {python_path} -m auto_round --model benzart/gemma-2b-it-fine-tuning-for-code-test "
+            f"cd ../.. && {python_path} -m auto_round --model /tf_dataset/auto_round/models/benzart/gemma-2b-it-fine-tuning-for-code-test "
             f" --bs 16 --iters 0 --nsamples 1 --format gguf:q4_k_m"
         )
         if res > 0 or res == -1:
@@ -54,8 +55,7 @@ class TestGGUF(unittest.TestCase):
     def test_q4_0(self):
         bits, group_size, sym = 4, 32, True
         autoround = AutoRound(
-            self.model,
-            self.tokenizer,
+            self.model_name,
             bits=bits,
             group_size=group_size,
             sym=sym,
@@ -102,8 +102,7 @@ class TestGGUF(unittest.TestCase):
     def test_func(self):
         bits, group_size, sym = 4, 128, True
         autoround = AutoRound(
-            self.model,
-            self.tokenizer,
+            self.model_name,
             # bits=bits,
             # group_size=group_size,
             # sym=sym,
@@ -196,7 +195,7 @@ class TestGGUF(unittest.TestCase):
     #     shutil.rmtree("./saved", ignore_errors=True)
 
     def test_gguf_baseline(self):
-        model_name = "Qwen/Qwen2.5-1.5B-Instruct"
+        model_name = "/tf_dataset/auto_round/models/Qwen/Qwen2.5-1.5B-Instruct"
         model = AutoModelForCausalLM.from_pretrained(model_name, torch_dtype="auto", trust_remote_code=True)
         autoround = AutoRound(
             model,
@@ -243,7 +242,7 @@ class TestGGUF(unittest.TestCase):
         # shutil.rmtree("./saved", ignore_errors=True)
 
     def test_q4_k_m(self):
-        model_name = "Qwen/Qwen2.5-1.5B-Instruct"
+        model_name = "/tf_dataset/auto_round/models/Qwen/Qwen2.5-1.5B-Instruct"
         model = AutoModelForCausalLM.from_pretrained(model_name, torch_dtype="auto", trust_remote_code=True)
         tokenizer = AutoTokenizer.from_pretrained(model_name, trust_remote_code=True)
         layer_config = {
@@ -290,7 +289,7 @@ class TestGGUF(unittest.TestCase):
         shutil.rmtree("./saved", ignore_errors=True)
 
     def test_all_format(self):
-        model_name = "Qwen/Qwen2.5-1.5B-Instruct"
+        model_name = "/tf_dataset/auto_round/models/Qwen/Qwen2.5-1.5B-Instruct"
         python_path = sys.executable
         # for gguf_format in ["gguf:q4_0", "gguf:q4_1", "gguf:q4_k_m", "gguf:q6_k"]:
         for gguf_format in ["gguf:q4_k_m"]:
@@ -311,8 +310,8 @@ class TestGGUF(unittest.TestCase):
             shutil.rmtree("../../tmp_autoround", ignore_errors=True)
 
     def test_vlm_gguf(self):
-        model_name = "Qwen/Qwen2-VL-2B-Instruct"
-        from auto_round.mllm.autoround_mllm import AutoRoundMLLM
+        model_name = "/tf_dataset/auto_round/models/Qwen/Qwen2-VL-2B-Instruct"
+        from auto_round import AutoRoundMLLM
         from auto_round.utils import mllm_load_model
 
         model, processor, tokenizer, image_processor = mllm_load_model(model_name)
@@ -328,13 +327,83 @@ class TestGGUF(unittest.TestCase):
         autoround.quantize_and_save(output_dir=quantized_model_path, format="gguf:q4_0")
         self.assertTrue("mmproj-model.gguf" in os.listdir("./saved"))
         for file_name in os.listdir(quantized_model_path):
-            file_name = os.listdir(quantized_model_path)[0]
             file_size = os.path.getsize(os.path.join(quantized_model_path, file_name)) / 1024**2
             if file_name == "mmproj-model.gguf":
-                self.assertAlmostEqual(file_size, 2535, delta=1.0)
+                self.assertAlmostEqual(file_size, 2537, delta=5.0)
             else:
-                self.assertAlmostEqual(file_size, 892, delta=1.0)
+                self.assertAlmostEqual(file_size, 892, delta=5.0)
         shutil.rmtree("./saved", ignore_errors=True)
+
+    def test_qtype_setting(self):
+        # Qwen2.5-0.5B-Instruct no output, token_embed q6_k fallbakc to q8_0 336M
+        # Qwen3-0.6B output q6_k, token_embed q4_0  448M
+        # Qwen3-8B output q6_k, token_embed q4_0 4.5G
+        # Llama-3.2-1B-Instruct o output, token_embed q6_k 736M
+        from auto_round.compressors.utils import set_layer_config
+        from auto_round.export.export_to_gguf.config import ModelType
+
+        model_name = "/tf_dataset/auto_round/models/Qwen/Qwen2.5-0.5B-Instruct"
+        ar = AutoRound(model=model_name, scheme="gguf:q4_0", iters=0)
+        ar.formats = ["gguf:q4_0"]
+        ar.layer_config, _, _ = set_layer_config(
+            ar.model,
+            ar.layer_config,
+            ar.scheme,
+            ar.scale_dtype,
+            ar.supported_types,
+            ar.inner_supported_types,
+            ar.quant_block_list,
+            ar.fp_layers,
+            ar.quant_lm_head,
+            enable_gguf_official_mixed=True,
+            is_mllm=ar.mllm,
+        )
+        self.assertTrue(ar.layer_config["model.embed_tokens"]["bits"] == 8)
+        self.assertTrue("lm_head" not in ar.layer_config)
+
+        model_name = "Qwen/Qwen3-0.6B"
+        ar = AutoRound(model=model_name, scheme="gguf:q4_0", iters=0)
+        ar.formats = ["gguf:q4_0"]
+        ar.layer_config, _, _ = set_layer_config(
+            ar.model,
+            ar.layer_config,
+            ar.scheme,
+            ar.scale_dtype,
+            ar.supported_types,
+            ar.inner_supported_types,
+            ar.quant_block_list,
+            ar.fp_layers,
+            ar.quant_lm_head,
+            enable_gguf_official_mixed=True,
+            is_mllm=ar.mllm,
+        )
+        self.assertTrue(ar.layer_config["model.embed_tokens"]["bits"] == 4)
+        self.assertTrue(ar.layer_config["lm_head"]["bits"] == 6 and ar.layer_config["lm_head"]["super_bits"] == 8)
+
+        layer_config = {
+            "model.embed_tokens": {"bits": 6, "super_bits": 8},
+            "lm_head": {"bits": 4},
+        }
+        ar = AutoRound(model=model_name, scheme="gguf:q4_0", iters=0, layer_config=layer_config)
+        ar.formats = ["gguf:q4_0"]
+        ar.layer_config, _, _ = set_layer_config(
+            ar.model,
+            ar.layer_config,
+            ar.scheme,
+            ar.scale_dtype,
+            ar.supported_types,
+            ar.inner_supported_types,
+            ar.quant_block_list,
+            ar.fp_layers,
+            ar.quant_lm_head,
+            enable_gguf_official_mixed=True,
+            is_mllm=ar.mllm,
+        )
+        self.assertTrue(ar.layer_config["lm_head"]["bits"] == 4)
+        self.assertTrue(
+            ar.layer_config["model.embed_tokens"]["bits"] == 6
+            and ar.layer_config["model.embed_tokens"]["super_bits"] == 8
+        )
 
 
 if __name__ == "__main__":
