@@ -38,7 +38,7 @@ class TestAutoRound(unittest.TestCase):
         shutil.rmtree("runs", ignore_errors=True)
 
     def test_load_gptq_no_dummy_gidx_model(self):
-        model_name = "ModelCloud/Llama-3.2-1B-Instruct-gptqmodel-4bit-vortex-v1"
+        model_name = "/tf_dataset/auto_round/models/ModelCloud/Llama-3.2-1B-Instruct-gptqmodel-4bit-vortex-v1"
         quantization_config = AutoRoundConfig()
         with self.assertRaises(NotImplementedError) as cm:
             model = AutoModelForCausalLM.from_pretrained(
@@ -50,7 +50,7 @@ class TestAutoRound(unittest.TestCase):
             )
 
     def test_load_awq(self):
-        model_name = "casperhansen/opt-125m-awq"
+        model_name = "/tf_dataset/auto_round/models/casperhansen/opt-125m-awq"
         quantization_config = AutoRoundConfig()
         model = AutoModelForCausalLM.from_pretrained(
             model_name,
