@@ -21,7 +21,7 @@ class LLMDataLoader:
 class TestAutoRound(unittest.TestCase):
     @classmethod
     def setUpClass(self):
-        self.model_name = "facebook/opt-125m"
+        self.model_name = "/tf_dataset/auto_round/models/facebook/opt-125m"
         self.save_folder = "./saved"
         self.llm_dataloader = LLMDataLoader()
 
@@ -31,7 +31,14 @@ class TestAutoRound(unittest.TestCase):
         shutil.rmtree("runs", ignore_errors=True)
 
     def test_gguf(self):
-        ar = AutoRound("Qwen/Qwen3-0.6B", scheme="W2A16", nsamples=1, iters=1, seqlen=2, dataset=self.llm_dataloader)
+        ar = AutoRound(
+            "/tf_dataset/auto_round/models/Qwen/Qwen3-0.6B",
+            scheme="W2A16",
+            nsamples=1,
+            iters=1,
+            seqlen=2,
+            dataset=self.llm_dataloader,
+        )
         ar.quantize_and_save(self.save_folder, format="gguf:q4_k_m")
         self.assertEqual(ar.bits, 4)
         shutil.rmtree(self.save_folder, ignore_errors=True)
@@ -57,7 +64,9 @@ class TestAutoRound(unittest.TestCase):
     def test_vllm(self):
         from auto_round import AutoRoundMLLM
 
-        ar = AutoRoundMLLM("Qwen/Qwen2-VL-2B-Instruct", scheme="W2A16", nsamples=1, iters=1, seqlen=2)
+        ar = AutoRoundMLLM(
+            "/tf_dataset/auto_round/models/Qwen/Qwen2-VL-2B-Instruct", scheme="W2A16", nsamples=1, iters=1, seqlen=2
+        )
         self.assertEqual(ar.bits, 2)
         self.assertEqual(ar.act_bits, 16)
 
@@ -69,6 +78,19 @@ class TestAutoRound(unittest.TestCase):
         self.assertEqual(ar.act_data_type, "nv_fp4_with_static_gs")
         ar.quantize()
 
+    def test_all_scheme(self):
+        import copy
+
+        preset_schemes = ["W8A16", "MXFP8", "FPW8A16", "FP8_STATIC", "GGUF:Q2_K_S", "GGUF:Q4_K_M"]
+        for scheme in preset_schemes:
+            model_name = self.model_name
+            if "gguf" in scheme.lower():
+                model_name = "/tf_dataset/auto_round/models/Qwen/Qwen2.5-1.5B-Instruct"
+            print(f"scheme={scheme}")
+            ar = AutoRound(model_name, scheme=scheme, nsamples=1, iters=1, seqlen=2, dataset=self.llm_dataloader)
+            ar.quantize_and_save(self.save_folder)
+            shutil.rmtree(self.save_folder, ignore_errors=True)
+
     def test_scheme_in_layer_config(self):
         layer_config = {
             "model.decoder.layers.2.self_attn": {"bits": 2},
@@ -76,7 +98,7 @@ class TestAutoRound(unittest.TestCase):
             "model.decoder.layers.4.self_attn.k_proj": QuantizationScheme.from_dict({"group_size": 64}),
         }
         ar = AutoRound(
-            self.model_name,
+            "/tf_dataset/auto_round/models/facebook/opt-125m",
             scheme="W3A16",
             nsamples=1,
             iters=1,
