@@ -9,7 +9,7 @@ import transformers
 from transformers import AutoConfig, AutoModelForCausalLM, AutoTokenizer
 
 from auto_round import AutoRound
-from auto_round.testing_utils import require_awq, require_optimum
+from auto_round.testing_utils import require_awq, require_optimum, require_package_version_ut
 
 
 class LLMDataLoader:
@@ -185,6 +185,7 @@ class TestAutoRound(unittest.TestCase):
         shutil.rmtree("./saved", ignore_errors=True)
 
     @require_awq
+    @require_package_version_ut("transformers", "<4.57.0")
     def test_autoawq_format(self):
         model = AutoModelForCausalLM.from_pretrained(self.model_name, torch_dtype="auto", trust_remote_code=True)
         tokenizer = AutoTokenizer.from_pretrained(self.model_name, trust_remote_code=True)
@@ -218,6 +219,7 @@ class TestAutoRound(unittest.TestCase):
 
     @require_optimum
     @require_awq
+    @require_package_version_ut("transformers", "<4.57.0")
     def test_autoawq_format_fp_qsave_layers(self):
         model = AutoModelForCausalLM.from_pretrained(self.model_name, torch_dtype="auto", trust_remote_code=True)
         layer_config = {
@@ -319,86 +321,6 @@ class TestAutoRound(unittest.TestCase):
         inputs = tokenizer(text, return_tensors="pt").to(model.device)
         print(tokenizer.decode(model.generate(**inputs, max_new_tokens=50)[0]))
         shutil.rmtree("./saved", ignore_errors=True)
-
-    def test_fp8input_mxfp4_llmcompressor_format(self):
-        model_name = "/models/Qwen3-0.6B-FP8"
-        scheme = "mxfp4"
-        ar = AutoRound(
-            model=model_name,
-            iters=2,
-            seqlen=2,
-            scheme=scheme,
-            dataset=self.llm_dataloader,
-        )
-        compressed_model, _ = ar.quantize_and_save(output_dir=self.save_dir, format="llm_compressor")
-        tmp_layer = compressed_model.model.layers[3].self_attn.q_proj
-        assert (
-            hasattr(tmp_layer, "weight_scale")
-            and hasattr(tmp_layer, "weight_packed")
-            and tmp_layer.weight_scale.dtype is torch.uint8
-            and tmp_layer.weight_scale.shape[0] == 2048
-        ), "Illegal MXFP4 packing name or data_type or shape"
-        quantization_config = AutoConfig.from_pretrained(self.save_dir, trust_remote_code=True).quantization_config
-        assert (
-            quantization_config["format"] == "float-quantized"
-            and quantization_config["config_groups"]["group_0"]["weights"]["is_mx"] is True
-            and quantization_config["config_groups"]["group_0"]["weights"]["num_bits"] == 4
-        ), f"Invalid MXFP4 quantization configuration: {quantization_config}"
-        shutil.rmtree(self.save_dir, ignore_errors=True)
-
-    def test_nvfp4_llmcompressor_format(self):
-        model = AutoModelForCausalLM.from_pretrained(self.model_name, torch_dtype="auto", trust_remote_code=True)
-        tokenizer = AutoTokenizer.from_pretrained(self.model_name, trust_remote_code=True)
-        scheme = "nvfp4"
-        autoround = AutoRound(
-            model,
-            tokenizer,
-            scheme=scheme,
-            iters=2,
-            seqlen=2,
-            dataset=self.llm_dataloader,
-        )
-        quantized_model_path = self.save_dir
-        compressed_model, _ = autoround.quantize_and_save(output_dir=quantized_model_path, format="llm_compressor")
-        tmp_layer = compressed_model.model.decoder.layers[3].self_attn.q_proj
-        assert (
-            hasattr(tmp_layer, "weight_scale")
-            and hasattr(tmp_layer, "weight_global_scale")
-            and hasattr(tmp_layer, "input_global_scale")
-            and tmp_layer.weight_packed.dtype is torch.uint8
-            and tmp_layer.weight_scale.dtype is torch.float8_e4m3fn
-            and tmp_layer.weight_scale.shape[0] == 768
-        ), "Illegal NVFP4 packing name or data_type or shape"
-        quantization_config = AutoConfig.from_pretrained(
-            quantized_model_path, trust_remote_code=True
-        ).quantization_config
-        assert (
-            quantization_config["format"] == "nvfp4-pack-quantized"
-            and quantization_config["config_groups"]["group_0"]["input_activations"]["num_bits"] == 4
-        ), f"Invalid NVFP4 quantization configuration: {quantization_config}"
-        shutil.rmtree("./saved", ignore_errors=True)
-        # from vllm import LLM, SamplingParams
-        # prompts = [
-        #     "The capital of France is",
-        #     "The future of AI is",
-        # ]
-        ## Create a sampling params object.
-        # sampling_params = SamplingParams(temperature=0.8, top_p=0.95, max_tokens=20)
-        # QUANTIZATION = "compressed-tensors"
-        # llm = LLM(model=quantized_model_path,
-        #         #   quantization=QUANTIZATION,
-        #           trust_remote_code=True,
-        #           tensor_parallel_size=1,
-        #           enforce_eager=True,
-        #           gpu_memory_utilization=0.7,
-        # )
-        # outputs = llm.generate(prompts, sampling_params)
-        # # Print the outputs.
-        # for output in outputs:
-        #     prompt = output.prompt
-        #     generated_text = output.outputs[0].text
-        #     if "France" in prompt:
-        #         assert "Paris" in generated_text
 
 
 if __name__ == "__main__":
