@@ -9,10 +9,11 @@ import torch
 import transformers
 from lm_eval.utils import make_table  # pylint: disable=E0401
 from transformers import AutoModelForCausalLM, AutoTokenizer
+from transformers.utils.versions import require_version
 
 from auto_round import AutoRound, AutoRoundAdam
 from auto_round.eval.evaluation import simple_evaluate
-from auto_round.testing_utils import require_awq, require_gptqmodel, require_optimum
+from auto_round.testing_utils import require_awq, require_gptqmodel, require_optimum, require_package_version_ut
 
 
 def get_accuracy(data):
@@ -38,7 +39,6 @@ class TestMainFunc(unittest.TestCase):
 
     @require_gptqmodel
     @require_optimum
-    @require_awq
     def test_backend(self):
         model_name = "/models/opt-125m"
         model = AutoModelForCausalLM.from_pretrained(model_name, torch_dtype=torch.float16, device_map="auto")
@@ -64,6 +64,16 @@ class TestMainFunc(unittest.TestCase):
         assert accuracy > 0.35
         shutil.rmtree("./saved", ignore_errors=True)
 
+    @require_optimum
+    @require_awq
+    @require_package_version_ut("transformers", "<4.57.0")
+    def test_backend_awq(self):
+        model_name = "/models/opt-125m"
+        model = AutoModelForCausalLM.from_pretrained(model_name, torch_dtype=torch.float16, device_map="auto")
+        tokenizer = AutoTokenizer.from_pretrained(model_name)
+        autoround = AutoRound(model, tokenizer, bits=4, group_size=128)
+        autoround.quantize()
+
         ##test auto_awq format
         autoround.save_quantized(self.save_dir, format="auto_awq", inplace=False)
         model_args = f"pretrained={self.save_dir}"
@@ -75,12 +85,11 @@ class TestMainFunc(unittest.TestCase):
 
     @unittest.skipIf(torch.cuda.is_available() is False, "Skipping because no cuda")
     @require_gptqmodel
-    @require_awq
     def test_fp_layers(self):
         model_name = "/models/opt-125m"
         model = AutoModelForCausalLM.from_pretrained(model_name, torch_dtype=torch.float16, device_map="auto")
         tokenizer = AutoTokenizer.from_pretrained(model_name)
-        from auto_round.utils import get_fp_layer_names
+        from auto_round.compressors.utils import get_fp_layer_names
 
         layer_names = get_fp_layer_names(model, "model.decoder.layers.0,model.decoder.layers.1")
         layer_configs = {}
@@ -97,6 +106,22 @@ class TestMainFunc(unittest.TestCase):
         accuracy = get_accuracy(res)
         assert accuracy > 0.35
         shutil.rmtree("./saved", ignore_errors=True)
+
+    @unittest.skipIf(torch.cuda.is_available() is False, "Skipping because no cuda")
+    @require_awq
+    @require_package_version_ut("transformers", "<4.57.0")
+    def test_fp_layers_awq(self):
+        model_name = "/models/opt-125m"
+        model = AutoModelForCausalLM.from_pretrained(model_name, torch_dtype=torch.float16, device_map="auto")
+        tokenizer = AutoTokenizer.from_pretrained(model_name)
+        from auto_round.compressors.utils import get_fp_layer_names
+
+        layer_names = get_fp_layer_names(model, "model.decoder.layers.0,model.decoder.layers.1")
+        layer_configs = {}
+        for name in layer_names:
+            layer_configs[name] = {"bits": 16}
+        autoround = AutoRound(model, tokenizer, bits=4, group_size=128)
+        autoround.quantize()
 
         ##test auto_awq format
         autoround.save_quantized(self.save_dir, format="auto_awq", inplace=False)
@@ -153,6 +178,26 @@ class TestMainFunc(unittest.TestCase):
         accuracy = get_accuracy(res)
         assert accuracy > 0.35
         shutil.rmtree("./saved", ignore_errors=True)
+
+    def test_attention_mask_lm_head(self):
+        from transformers import AutoTokenizer
+
+        model_name = "/models/Qwen3-8B"
+        # model_name = "/models/Qwen3-0.6B"
+        tokenizer = AutoTokenizer.from_pretrained(model_name)
+        text = ["haha", "hello world"]
+        res = tokenizer(text, return_tensors="pt", max_length=8, padding="max_length", truncation=True)
+        res.data.pop("attention_mask")
+        data = [res.data]
+
+        text = ["qudd", "hfd"]
+        res = tokenizer(text, return_tensors="pt", max_length=8, padding="max_length", truncation=True)
+        res.data.pop("attention_mask")
+        data.append(res.data)
+        from auto_round import AutoRound
+
+        ar = AutoRound(model_name, iters=1, dataset=data, seqlen=8, quant_lm_head=True)
+        ar.quantize()
 
 
 if __name__ == "__main__":
